Refactor to extract XPaths to a separate file
